removed calls to String from BigInt marshalling unmarshalling
